Adds he/him (femme clothes) and she/her (male clothes) pronouns
Updated gender checks for female/male clothes for the roles that use them, now you can have she/her pronouns and have guy clothes and vice versa.

This was a lot of roles to edit and I might have overlooked some. If I did, post it and I'll fix it.