Swift Journey
After speaking with one of the devs I brought to their attention that the Swift Journey, The Necran Flail, That is one of the starting chosen special themed church weapons that the templars can choose from, but for some reason it is the only weapon from all these special weapons that is based on it's IRON counterpart not it's steel counterpart. After getting approval I put the weapon in line with the other steel god weapons.