Normalized Skirt Length, Plate Tassets
Extended the described and sprited length of the chain skirt to match that of the normal skirt, which was described as 'modest'. 

Made the chain skirt cover legs. Why the armored version was only groin covering is beyond me; we have the brayette for that.

Removed the 'mini-skirt' from the plate skirt; now described as tassets.

Tassets, like the skirt, cover the upper thigh, hence leg armor. This is where the most muscle and arteries are exposed to attack, compared to the shin which is a stretch for most attack geometry and not a suitable target.