implements slapcrafting as a component, refactors a bunch of shit to support it plus precursors for future refactors (#2294)
* implements slapcrafting as a component

implements slapcrafting as a component so that we can eventually blow up fragile and difficult to modify and maintain if stacks for slapcrafting recipes.

in the process of doing this ive had to overhaul a lot of other secondary things to bring them closer to contemporary tg code standards. most of this work is done, and this build compiles, but it produces a significant amount of runtimes based around about four broken things right now.

* some more edits. trust the plan.

* single line change

* clears runtime + some storage improvements

* solves a major runtime destructively

* runtime fixes

* more runtime clears

* nukes atom_storage for not workin out

* ruby is no longer the base gem, datumizes staffcrafting

* Update dun_world.dmm

* blow me dun world

* Revert "blow me dun world"

This reverts commit a490566256c8b7f9b85d7f87154a5f96f721736b.

* Update dun_world.dmm

* sdmm blow me good and hard

* Revert "sdmm blow me good and hard"

This reverts commit 427ae91f661b19ad0a9a56f57079efb0c231b11d.

* Revert "Merge branch 'slapcraft-blow-me-good-and-hard-through-my-jorts' of https://github.com/kittysmooch/Azure-Peak into slapcraft-blow-me-good-and-hard-through-my-jorts"

This reverts commit 3ac57e8a385825ac141ec90088c55bc45a9681dd, reversing
changes made to 427ae91f661b19ad0a9a56f57079efb0c231b11d.

* Reapply "Merge branch 'slapcraft-blow-me-good-and-hard-through-my-jorts' of https://github.com/kittysmooch/Azure-Peak into slapcraft-blow-me-good-and-hard-through-my-jorts"

This reverts commit c4a97915065697117e73088f9f1a05e047a90c83.

* kill all mappers NOW