Skirt and tasset leg coverage.
Following a concept of 'reduced coverage for higher protection', shifted skirts to be chainmail leg armor with the groin as a weak spot. Allows light and medium armor users to 'step up' their leg armor without losing dodge or requiring heavy armor training, but simultaneously creates a new risk of groin strikes. This in turn shifts attacks to the chest, creating a shift in the combat meta away from limb strikes.

Discussed with Neo directly.