The return of the Duke's armor
In the past, the Duke started out with a set of Blacksteel armor. However, I honest to god found Blacksteel to be so buttfuck ugly that I removed it from the Duke's armoire. I could not find it in myself to enjoy or even remotely like the armor's design. It was purely an aesthetic choice that I had made.

But now. Now that my glorious man Dakken has gotten us the new modern Blacksteel armor sprites, I am replacing the armor from the Duke's armoire and returning what had once been taken.

This does not mean that duke players suddenly have the excuse to go and be frontliners. It isn't an right to frag.

This will be the first of a good few map changes I intend to make.