Fixes
Changed the sound to the deadbell then a loud ass churn felt more thematic..I could have also found some creepy whispers.

IcOn MeRge ConFlicts

added a furhter check for the dispelled logic.

Dystroyed check for if anything BESIDES THE CASTER removes the portal from existance.

The portal now respawns correctly it needed to be if(src in contents) not if(loc in contents) I had initally thought that as loc is checking for the existance of a location not an object in a thing. This is the one cause the implied src needs to be FURTHER implied.