Difficulty and Reward tweaking
We call this a difficulty tweak!

Tweaks the rewards of various dungeons. The orc fortress dungeon has been made a fair bit more difficult, and some of its rewards were tweaked. May need to testmerge and see how it really is. Also, I'm not sorry staff fiends.

Switched out the coin spawners in the church vault to be a bit more consistent to a degree.