Old god patron change
Gives psydonites the hidden patron option, which for coded purposes, gives them the same role list as inhumen worshippers, as they are making an effort to hide their faith. Only ten worshippers are "allowed" to be nobility by the holy see.
Open psydonites will have a smaller role list but will be tolerated.